refactor(schemaToForm): split between SchemaComponents & subfunctions
