+++ app/components/form/SchemaComponents.tsx
@@ -1,19 +1,29 @@
 import type { PageSchema } from "~/domains/pageConfig";
 import type { StrapiFormComponent } from "~/services/cms/models/StrapiFormComponent";
-import { schemaToFormElement } from "./schemaToForm";
+import { getNestedSchema } from "./schemaToForm/getNestedSchema";
+import { isZodEnum, renderZodEnum } from "./schemaToForm/renderZodEnum";
+import { isZodString, renderZodString } from "./schemaToForm/renderZodString";
 
-export const SchemaComponents = ({
-  pageSchema,
-  formElements,
-}: {
+type Props = {
   pageSchema: PageSchema;
   formElements?: StrapiFormComponent[];
-}) => {
+};
+
+export const SchemaComponents = ({ pageSchema, formElements }: Props) => {
   return (
     <div className="ds-stack ds-stack-40">
-      {Object.entries(pageSchema).map(([fieldName, fieldSchema]) =>
-        schemaToFormElement(fieldName, fieldSchema, formElements),
-      )}
+      {Object.entries(pageSchema).map(([fieldName, fieldSchema]) => {
+        const nestedSchema = getNestedSchema(fieldSchema);
+        const matchingElement = formElements?.find(
+          ({ name }) => name === fieldName,
+        );
+
+        if (isZodEnum(nestedSchema))
+          return renderZodEnum(nestedSchema, fieldName, matchingElement);
+
+        if (isZodString(nestedSchema))
+          return renderZodString(nestedSchema, fieldName, matchingElement);
+      })}
     </div>
   );
 };
